Cleanup and fix Persona search.
